Correcting grammar just a bit
But a bit more explanation would be needed @mariosky. Besides, it's
not clear to me what is the mapping from the parameter tables to these
results. Are we doing these charts for every possible combination? Or
are we simply using different parameters for different number of
dimensions?

Besides, There seem to be some cases where our results are better than
the baseline (diamonds): Skew Rastrigin-Bueche, for instance. Is that
the case? Or is it only for a certain amount of effort?